mostly refactored UserCheckboxList
Doesn't send its data or interact upwards yet
