+++ src/components/UserCheckboxList.js
@@ -1,161 +1,34 @@
-import React from 'react';
+/* eslint-disable react/prefer-stateless-function */
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import '../css/GenreList.css';
+import { FontAwesomeIcon } from '../utils';
+import '../css/UserCheckboxList.css';
 
-// Renders the genre list to the MovieList page. ==================
-export default class UserCheckboxList extends React.Component {
+export default class UserCheckboxList extends Component {
   static propTypes = {
-    genres: PropTypes.arrayOf(PropTypes.object).isRequired,
-    moviesByGenreId: PropTypes.func.isRequired,
-    checkboxesVisible: PropTypes.bool,
-    style: PropTypes.shape({
-      display: PropTypes.string,
-    }),
-    showCertifications: PropTypes.bool,
-    selectedGenres: PropTypes.arrayOf(PropTypes.string),
-    selectedCertifications: PropTypes.arrayOf(PropTypes.string),
-    handleFormChange: PropTypes.func,
+    options: PropTypes.arrayOf(PropTypes.string).isRequired,
   };
-
-  static defaultProps = {
-    style: 'style',
-    checkboxesVisible: false,
-    selectedGenres: [],
-    selectedCertifications: [],
-    showCertifications: false,
-    handleFormChange: null,
-  };
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      activeButton: '',
-    };
-
-    this.handleSelectionReset = this.handleSelectionReset.bind(this);
-    this.handlePreferencesChange = this.handlePreferencesChange.bind(this);
-  }
-
-  handleActiveButton = genreName => {
-    if (genreName) {
-      this.setState({
-        activeButton: genreName,
-      });
-    } else {
-      this.setState({
-        activeButton: '',
-      });
-    }
-  };
-
-  handlePreferencesChange(e, reset) {
-    const { handleFormChange } = this.props;
-    handleFormChange(e, reset);
-  }
-
-  // =================== Unticks all checked genreboxes ==============
-
-  handleSelectionReset(e) {
-    e.preventDefault();
-    e.target.reset();
-    this.handlePreferencesChange(e, true);
-  }
 
   render() {
-    const {
-      moviesByGenreId,
-      genres,
-      style,
-      checkboxesVisible,
-      showCertifications,
-      selectedGenres,
-      selectedCertifications,
-    } = this.props;
-    const { activeButton } = this.state;
-    const certifications = ['G', 'PG', 'PG-13', 'R', 'NC-17'];
-
-    if (showCertifications) {
-      return (
-        <div>
-          <form onSubmit={this.handleSelectionReset} id="certificationsForm">
-            <div id="list-genres">
-              {certifications.map(certification => (
-                <label className="single-genre checkmark-container" key={certification}>
-                  <input
-                    type="checkbox"
-                    name="certification"
-                    defaultChecked={selectedCertifications.includes(certification)}
-                    value={certification}
-                    onChange={this.handlePreferencesChange}
-                  />
-                  <span className="checkmark" />
-                  <span className="single-genre">{certification}</span>
-                </label>
-              ))}
-            </div>
-            <button className="reset-button" type="submit">
-              reset
-            </button>
-          </form>
-        </div>
-      );
-    }
-    if (checkboxesVisible) {
-      return (
-        <div>
-          <form onSubmit={this.handleSelectionReset} id="genrePreferencesForm">
-            <div id="list-genres">
-              {genres.map(genre => (
-                <label className="single-genre checkmark-container" key={genre._id}>
-                  <input
-                    type="checkbox"
-                    name="genre"
-                    value={genre._id}
-                    defaultChecked={selectedGenres.includes(genre._id)}
-                    onChange={this.handlePreferencesChange}
-                  />
-                  <span className="checkmark" />
-                  <span className="single-genre">{genre.translation}</span>
-                </label>
-              ))}
-            </div>
-            <button className="reset-button" type="submit">
-              reset
-            </button>
-          </form>
-        </div>
-      );
-    }
+    const { options } = this.props;
 
     return (
-      <div id="list-genres" style={style}>
-        <button
-          className="single-genre"
-          type="button"
-          key="all"
-          onClick={() => moviesByGenreId('all')}
-        >
-          All
-        </button>
-        {genres.map(genre => (
-          <button
-            className="single-genre"
-            type="button"
-            key={genre.technical_name}
-            style={
-              activeButton === genre.translation
-                ? { backgroundColor: '#88388c' }
-                : { backgroundColor: 'rgba(226, 217, 217, 0.0)' }
-            }
-            onClick={() => {
-              moviesByGenreId(genre._id);
-              this.handleActiveButton(genre.translation);
-            }}
-          >
-            {genre.translation}
-          </button>
-        ))}
-      </div>
+      <form className="checkbox-form">
+        <div className="checkboxes">
+          {options.map(option => (
+            <span className="single-option" key={option}>
+              <input className="checkbox" type="checkbox" id={option} value={option} />
+              <label className="option-label" htmlFor={option}>
+                {option}
+              </label>
+              <label className="check-tick" htmlFor={option}>
+                <FontAwesomeIcon icon="check" />
+              </label>
+            </span>
+          ))}
+        </div>
+        <input type="reset" className="reset" value="Reset" />
+      </form>
     );
   }
 }
