+++ src/css/GenreList.css
@@ -7,17 +7,6 @@
   background-color: rgba(226, 217, 217, 0.3);
 }
 
-#list-genres {
-  /* display: none; */
-  /* width: 100%;
-  left: 0;
-  background-color: var(--offwhite);
-  opacity: .1;
-  /* box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2); */
-  /* z-index: 1;
-  border-radius: 5px; */
-}
-
 #list-genres button {
   margin: 10px;
   border: none;
@@ -51,7 +40,6 @@
   transition: 180ms;
 }
 
-
 .single-genre {
   color: white;
 }
@@ -62,6 +50,6 @@
 
 @media only screen and (max-width: 600px) {
   #list-genres button {
-    font-size: .7rem;
+    font-size: 0.7rem;
   }
 }
