+++ src/css/UserCheckboxList.css
@@ -1,0 +1,104 @@
+@import './style-variables.css';
+
+.checkbox-form {
+  --edge-curve: 10px;
+
+  border-radius: var(--edge-curve);
+  background-color: rgba(226, 217, 217, 0.3);
+  color: var(--offwhite);
+  font-size: 1.1rem;
+}
+
+/* ======================================================= */
+
+.checkboxes {
+  --hover-color: var(--popcornyellow);
+
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+/* ======================================================= */
+
+.single-option {
+  position: relative;
+  display: flex;
+  align-items: center;
+  padding: 8px 12px;
+}
+
+.single-option .check-tick {
+  cursor: pointer;
+  position: absolute;
+  font-weight: bold;
+  left: 16px;
+  top: 17px;
+}
+
+.single-option .option-label {
+  cursor: pointer;
+  padding: 8px 0px;
+}
+
+.single-option:hover .check-tick {
+  color: var(--hover-color);
+}
+
+.single-option:hover .option-label {
+  text-shadow: 0.8px 0.8px 0.5px var(--red);
+  color: var(--hover-color);
+}
+
+.single-option:hover .checkbox {
+  background-color: var(--hover-color);
+}
+
+.single-option .checkbox {
+  appearance: none;
+  cursor: pointer;
+  height: 25px;
+  width: 25px;
+  background-color: var(--offwhite);
+  border: 1px solid var(--darkgrey);
+  border-radius: 3px;
+  margin-right: 4px;
+}
+
+.single-option .checkbox:checked {
+  background-color: var(--purple);
+}
+
+.single-option .checkbox:focus {
+  outline: none;
+}
+
+/* ======================================================= */
+
+.checkbox-form .reset {
+  color: var(--offwhite);
+  width: 100%;
+  padding: 5px 10px;
+  border: none;
+  border-top: 0.5px solid var(--darkgrey);
+  background-color: var(--purple);
+  border-radius: 0px 0px var(--edge-curve) var(--edge-curve);
+  font-size: 1em;
+}
+
+.checkbox-form .reset:active {
+  background-color: var(--red);
+}
+
+.checkbox-form .reset:hover {
+  cursor: pointer;
+  opacity: 0.8;
+  box-shadow: 0px 1px 1px 1px var(--red);
+}
+
+.checkbox-form .reset:focus {
+  outline: none;
+}
+
+/* ======================================================= */
